+++ src/Magento/MagentoCloud/Config/Deploy.php
@@ -5,10 +5,9 @@
  */
 namespace Magento\MagentoCloud\Config;
 
+use Magento\MagentoCloud\Config\Deploy\Reader;
+use Magento\MagentoCloud\Config\Deploy\Writer;
 use Magento\MagentoCloud\DB\ConnectionInterface;
-use Magento\MagentoCloud\Filesystem\DirectoryList;
-use Magento\MagentoCloud\Filesystem\Driver\File;
-use Magento\MagentoCloud\Filesystem\FileSystemException;
 use Psr\Log\LoggerInterface;
 
 /**
@@ -27,31 +26,31 @@ class Deploy
     private $connection;
 
     /**
-     * @var File
+     * @var Reader
      */
-    private $file;
+    private $reader;
 
     /**
-     * @var DirectoryList
+     * @var Writer
      */
-    private $directoryList;
+    private $writer;
 
     /**
      * @param LoggerInterface $logger
      * @param ConnectionInterface $connection
-     * @param File $file
-     * @param DirectoryList $directoryList
+     * @param Reader $reader
+     * @param Writer $writer
      */
     public function __construct(
         LoggerInterface $logger,
         ConnectionInterface $connection,
-        File $file,
-        DirectoryList $directoryList
+        Reader $reader,
+        Writer $writer
     ) {
         $this->logger = $logger;
         $this->connection = $connection;
-        $this->file = $file;
-        $this->directoryList = $directoryList;
+        $this->reader = $reader;
+        $this->writer = $writer;
     }
 
     /**
@@ -78,74 +77,15 @@ class Deploy
             throw new \Exception('Missing either core_config_data or setup_module table', 5);
         }
 
-        try {
-            $data = $this->getConfig();
-            if (isset($data['install']['date'])) {
-                $this->logger->info('Magento was installed on ' . $data['install']['date']);
-            }
-        } catch (FileSystemException $e) {
-            $this->write(['install' => ['date' => date('r')]]);
+        $data = $this->reader->read();
+        if (isset($data['install']['date'])) {
+            $this->logger->info('Magento was installed on ' . $data['install']['date']);
+
+            return true;
         }
+
+        $this->writer->write(['install' => ['date' => date('r')]]);
 
         return true;
-    }
-
-    /**
-     * Return full path to environment configuration file.
-     *
-     * @return string The path to configuration file
-     */
-    public function getConfigFilePath(): string
-    {
-        return $this->directoryList->getMagentoRoot() . '/app/etc/env.php';
-    }
-
-    /**
-     * Returns environment configuration.
-     *
-     * @return array
-     * @throws FileSystemException If configuration file not found
-     */
-    public function getConfig()
-    {
-        $configPath = $this->getConfigFilePath();
-        if (!$this->file->isExists($configPath)) {
-            throw new FileSystemException(
-                sprintf('Configuration file: %s not exists', $configPath)
-            );
-        }
-
-        return include $configPath;
-    }
-
-    /**
-     * Updates existence configuration.
-     *
-     * @param array $config
-     */
-    public function update(array $config)
-    {
-        try {
-            $oldConfig = $this->getConfig();
-        } catch (FileSystemException $e) {
-            $oldConfig = [];
-        }
-
-        $updatedConfig = array_replace_recursive($oldConfig, $config);
-        $updatedConfig = '<?php' . PHP_EOL . 'return ' . var_export($updatedConfig, true) . ';';
-
-        $this->file->filePutContents($this->getConfigFilePath(), $updatedConfig);
-    }
-
-    /**
-     * Writes given configuration to file.
-     *
-     * @param array $config
-     */
-    public function write(array $config)
-    {
-        $updatedConfig = '<?php' . PHP_EOL . 'return ' . var_export($config, true) . ';';
-
-        $this->file->filePutContents($this->getConfigFilePath(), $updatedConfig);
     }
 }
